[Feature] Add scraper for clinical officers, change folder structure
