+++ config.py
@@ -17,7 +17,7 @@ CLOUDSEARCH_DOCTORS_ENDPOINT = {
 
 # Clinical document endpoint
 CLOUDSEARCH_COS_ENDPOINT = {
-    "endpoint_url" : "http://doc-cfa-healthtools-ke-doctors-m34xee6byjmzcgzmovevkjpffy.eu-west-1.cloudsearch.amazonaws.com/",
+    "endpoint_url" : "http://doc-cfa-healthtools-ke-cos-nhxtw3w5goufkzram4er7sciz4.eu-west-1.cloudsearch.amazonaws.com/",
     "aws_access_key_id": os.getenv("MORPH_AWS_ACCESS_KEY_ID"),
     "aws_secret_access_key": os.getenv("MORPH_AWS_SECRET_KEY"),
     "region_name": os.getenv("MORPH_AWS_REGION")
