+++ tests/dummy_files/clinical_officers.json
@@ -1,0 +1,1 @@
+[{"fields": {"name": "JACOB KILIMO KISANG", "reg_no": "1 / Ra00302/17", "valid_dates": "2023-01-17 / 2031-01-19", "qualifications": "DIP", "reg_date": "1989-01-05T00:00:00.000Z", "address": "PO BOX 1881", "id": 0}, "type": "add", "id": 0}, {"fields": {"name": "FRANCIS KINGE MATHERI", "reg_no": "18 / Ra01477/16", "valid_dates": "2030-03-16 / 2031-03-18", "qualifications": "DIP", "reg_date": "1989-07-08T00:00:00.000Z", "address": "P.O BOX 656  KIAMBU", "id": 1}, "type": "add", "id": 1}, {"fields": {"name": "GLADYS OKAKAH KOYENGO", "reg_no": "19 / Ra00811/15", "valid_dates": "2018-11-15 / 2030-11-17", "qualifications": "DIP", "reg_date": "1989-04-26T00:00:00.000Z", "address": "P.O BOX 20715  NAIROBI", "id": 2}, "type": "add", "id": 2}, {"fields": {"name": "STEPHEN OKEYO ABEBE", "reg_no": "20 / R02062/16", "valid_dates": "2018-08-16 / 2031-08-18", "qualifications": "DIP", "reg_date": "1989-03-16T00:00:00.000Z", "address": "P.O. BOX 1  KADONGO", "id": 3}, "type": "add", "id": 3}, {"fields": {"name": "JOSEPH OGEKA MAORE", "reg_no": "30 / Ra01287/16", "valid_dates": "2014-03-16 / 2031-03-18", "qualifications": "DIP", "reg_date": "1989-01-05T00:00:00.000Z", "address": "P.O BOX 153  KISII", "id": 4}, "type": "add", "id": 4}, {"fields": {"name": "HAMISI ALI FAJIN", "reg_no": "37 / Ra01032/15", "valid_dates": "2008-12-15 / 2031-12-17", "qualifications": "DIP", "reg_date": "1989-01-10T00:00:00.000Z", "address": "P.O BOX 57  MALINDI", "id": 5}, "type": "add", "id": 5}, {"fields": {"name": "GEORGE AYIEKO JOWI", "reg_no": "40 / Ra01823/16", "valid_dates": "2012-05-16 / 2031-05-18", "qualifications": "DIP", "reg_date": "1989-01-05T00:00:00.000Z", "address": "P.O. BOX 538  MTWAPA", "id": 6}, "type": "add", "id": 6}, {"fields": {"name": "STEPHEN KIPLANGAT KITTUR", "reg_no": "48 / R02318/15", "valid_dates": "2013-04-15 / 2030-04-17", "qualifications": "DIP", "reg_date": "1987-08-25T00:00:00.000Z", "address": "P.O BOX 5036  ELDORET", "id": 7}, "type": "add", "id": 7}, {"fields": {"name": "ROBERT MUNGE KIMUNDUU", "reg_no": "54 / R02261/15", "valid_dates": "2013-04-15 / 2030-04-17", "qualifications": "DIP", "reg_date": "1988-02-02T00:00:00.000Z", "address": "P.O BOX 304  MACHAKOS", "id": 8}, "type": "add", "id": 8}, {"fields": {"name": "CHRISANTHUS MAKORI ORINA", "reg_no": "55 / Ra00375/15", "valid_dates": "2029-09-15 / 2030-09-17", "qualifications": "DIP", "reg_date": "1988-01-28T00:00:00.000Z", "address": "P.O. BOX 3685  KISII", "id": 9}, "type": "add", "id": 9}, {"fields": {"name": "EZEKIEL WALUFU WALUFU", "reg_no": "76 / R02849/15", "valid_dates": "2021-04-15 / 2030-04-17", "qualifications": "DIP", "reg_date": "1987-03-19T00:00:00.000Z", "address": "P.O BOX 802  MUMIAS", "id": 10}, "type": "add", "id": 10}, {"fields": {"name": "ELIKANAH KEBAGENDI OMWENGA", "reg_no": "78 / Ra00147/15", "valid_dates": "2001-09-15 / 2030-09-17", "qualifications": "DIP", "reg_date": "1989-08-16T00:00:00.000Z", "address": "P.O BOX 252 80100 MOMBASA", "id": 11}, "type": "add", "id": 11}, {"fields": {"name": "DANIEL OMBACHI MOGENI", "reg_no": "97 / Ra00643/16", "valid_dates": "2003-02-16 / 2028-02-18", "qualifications": "DIP", "reg_date": "1988-01-28T00:00:00.000Z", "address": "P.O. BOX 235  NYAMIRA", "id": 12}, "type": "add", "id": 12}, {"fields": {"name": "BENSON GICHUHI NG\\'ERI", "reg_no": "107 / Ra00293/15", "valid_dates": "2015-09-15 / 2030-09-17", "qualifications": "DIP", "reg_date": "1989-01-11T00:00:00.000Z", "address": "P.O. BOX 15383  NAKURU", "id": 13}, "type": "add", "id": 13}, {"fields": {"name": "ANDIVA HERMAN KAGHALI", "reg_no": "112 / Ra00758/16", "valid_dates": "2009-02-16 / 2028-02-18", "qualifications": "DIP", "reg_date": "1989-05-12T00:00:00.000Z", "address": "P.O. BOX 119 50309 TIRIKI", "id": 14}, "type": "add", "id": 14}, {"fields": {"name": "DABASO GARAMBODA JILLO", "reg_no": "121 / Ra00245/17", "valid_dates": "2019-01-17 / 2031-01-19", "qualifications": "DIP", "reg_date": "1989-01-17T00:00:00.000Z", "address": "P.O BOX 753  GARISSA", "id": 15}, "type": "add", "id": 15}, {"fields": {"name": "SIYAD MOHAMED OMAR", "reg_no": "122 / Ra02079/16", "valid_dates": "2015-06-16 / 2030-06-18", "qualifications": "DIP", "reg_date": "1989-01-06T00:00:00.000Z", "address": "P.O BOX 253 GARISSA", "id": 16}, "type": "add", "id": 16}, {"fields": {"name": "DAVID MAINA NJURU", "reg_no": "125 / Ra02390/16", "valid_dates": "2019-07-16 / 2031-07-18", "qualifications": "DIP", "reg_date": "1989-10-12T00:00:00.000Z", "address": "P.O BOX 20435 00200 NAIROBI", "id": 17}, "type": "add", "id": 17}, {"fields": {"name": "ALLEN PETERSON MBELA", "reg_no": "138 / Ra00187/16", "valid_dates": "2013-01-16 / 2031-01-18", "qualifications": "DIP", "reg_date": "1989-01-17T00:00:00.000Z", "address": "P.O. BOX 47 MWATATE", "id": 18}, "type": "add", "id": 18}, {"fields": {"name": "LINCOLN MURAI NJUGUNA", "reg_no": "139 / Ra00572/15", "valid_dates": "2001-11-15 / 2030-11-17", "qualifications": "DIP", "reg_date": "1988-02-12T00:00:00.000Z", "address": "P.O BOX 125  KILGORIS", "id": 19}, "type": "add", "id": 19}, {"fields": {"name": "KITHUSI PETER KIITI", "reg_no": "141 / Ra01441/16", "valid_dates": "2024-03-16 / 2031-03-18", "qualifications": "DIP", "reg_date": "1989-02-21T00:00:00.000Z", "address": "P.O BOX 1473  KANGUNDO", "id": 20}, "type": "add", "id": 20}, {"fields": {"name": "MICHAEL GITARI KAMAU", "reg_no": "147 / R03501/15", "valid_dates": "2016-07-15 / 2031-07-17", "qualifications": "DIP", "reg_date": "1989-01-20T00:00:00.000Z", "address": "P.O BOX 701  THIKA", "id": 21}, "type": "add", "id": 21}, {"fields": {"name": "JAMES NYORO WAMWEA", "reg_no": "163 / Ra01474/16", "valid_dates": "2029-03-16 / 2031-03-18", "qualifications": "DIP", "reg_date": "2055-03-15T00:00:00.000Z", "address": "P.O BOX 265  LIMURU", "id": 22}, "type": "add", "id": 22}, {"fields": {"name": "GITHAIGA NJUMA ONESMUS", "reg_no": "168 / Ra01203/16", "valid_dates": "2008-03-16 / 2031-03-18", "qualifications": "DIP", "reg_date": "1989-01-25T00:00:00.000Z", "address": "P.O. BOX 171  MUKURWEINI", "id": 23}, "type": "add", "id": 23}, {"fields": {"name": "JUSTUS KARIGU KIBARUA", "reg_no": "171 / Ra01199/16", "valid_dates": "2008-03-16 / 2031-03-18", "qualifications": "DIP", "reg_date": "2001-01-31T10:46:00.000Z", "address": "P.O. BOX 2  GIAKANJA", "id": 24}, "type": "add", "id": 24}, {"fields": {"name": "KILLIAN ANSELM MWOLOI", "reg_no": "178 / Ra00200/16", "valid_dates": "2013-01-16 / 2031-01-18", "qualifications": "DIP", "reg_date": "1989-02-21T00:00:00.000Z", "address": "P.O. BOX 152  KIBWEZI", "id": 25}, "type": "add", "id": 25}, {"fields": {"name": "CHARLES GATHUMA BIRINGI", "reg_no": "182 / Ra01202/16", "valid_dates": "2008-03-16 / 2031-03-18", "qualifications": "DIP", "reg_date": "2011-03-15T11:34:00.000Z", "address": "P.O BOX 920  NYERI", "id": 26}, "type": "add", "id": 26}, {"fields": {"name": "ANTONY KITSAO CHIRAHU", "reg_no": "198 / Ra00540/15", "valid_dates": "2021-10-15 / 2031-10-17", "qualifications": "DIP", "reg_date": "1989-01-31T00:00:00.000Z", "address": "P.O BOX  49  MOMBASA", "id": 27}, "type": "add", "id": 27}, {"fields": {"name": "JULIUS MUTAMBUKI KILIKU", "reg_no": "206 / Ra01512/16", "valid_dates": "2001-05-16 / 2031-05-18", "qualifications": "DIP", "reg_date": "2011-03-17T14:20:00.000Z", "address": "P.O BOX 350  90119 MATUU", "id": 28}, "type": "add", "id": 28}, {"fields": {"name": "STEPHEN  ODUOR JURE", "reg_no": "207 / Ra00326/15", "valid_dates": "2018-09-15 / 2030-09-17", "qualifications": "DIP", "reg_date": "1989-07-12T00:00:00.000Z", "address": "P.O BOX 370  KISUMU", "id": 29}, "type": "add", "id": 29}]
